+++ CSharp/libCpScript.Net.Asm/State.cs
@@ -106,15 +106,15 @@ namespace libCpScript.Net.Asm
 			return ReturnValue;
 		}
 		
-		public int PopInt ()
+		public long PopInt ()
 		{
 			return NativeState.State_PopInt (_State);
 		}
 		
-		public int GetIntVariableInScope (string Name)
+		public long GetIntVariableInScope (string Name)
 		{
 			IntPtr AnsiName = Marshal.StringToHGlobalAnsi(Name);
-			int r = NativeState.State_GetIntVariableInScope(_State, AnsiName);
+			long r = NativeState.State_GetIntVariableInScope(_State, AnsiName);
 			Marshal.FreeHGlobal(AnsiName);
 			return r;
 		}
@@ -170,12 +170,12 @@ namespace libCpScript.Net.Asm
 			NativeState.State_Push(_State, GCHandle.ToIntPtr(tmpHandle));
 		}
 		
-		public void PushInt (int v)
+		public void PushInt (long v)
 		{
 			NativeState.State_PushInt(_State, v);
 		}
 		
-		public void SetIntVariableInScope (string Name, int v)
+		public void SetIntVariableInScope (string Name, long v)
 		{
 			IntPtr AnsiName = Marshal.StringToHGlobalAnsi (Name);
 			NativeState.State_SetIntVariableInScope(_State, AnsiName, v);
